moved landing page js moved to it's own file within components. I also added a style page specifically for the landing page styles
