+++ client/src/components/layout/Style.scss
@@ -55,27 +55,7 @@ $enable-gradients: true;
     }
     .search {
         padding: 15px;
-    }
-
-//landing page
-.landing-page{
-    padding: 0 40px;
-    h1 {
-        text-align: center;
-        color: $untGreen;
-        font-weight: 800;
-    }
-    /* Image */
-    img.center {
-        display: block;
-        margin: 0 auto;
-    }
-}
-
-#landingImg{
-    width: 30vw;
-}
-    
+    }   
 
 /* Button */
 .btn-primary {
